# src/main/java/com/example/demo/infrastructure/adapter/inbound/dto/ErrorResponse.java
package com.example.demo.infrastructure.adapter.inbound.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import java.time.LocalDateTime;
import lombok.Builder;
import lombok.Data;

@Data
@Builder
@Schema(description = "Objeto de error")
public class ErrorResponse {
  private LocalDateTime timestamp;
  private int status;
  private String error;
  private String message;
  private String path;
}